refactor: move extract data methods to a separate file
