+++ src/helpers/extractData/extractProductDataFromProduct.ts
@@ -1,0 +1,146 @@
+import type { Cart, Product } from '@commercetools/platform-sdk'
+import type { TCurrencyType, TLangType } from '@/types/appSettingsTypes'
+import { firstLetterUppercase } from '@/helpers/transformation/stringTransform'
+
+export function extractInCartNumber(product: Product, cart: Cart) {
+  const items = cart.lineItems
+
+  if (items) {
+    const lineItem = items.find((item) => item.productId === product.id)
+
+    if (lineItem) {
+      return lineItem.quantity
+    }
+    return 0
+  }
+  return 0
+}
+
+export function extractLineItemId(productId: string, cart: Cart) {
+  if (cart.lineItems.length !== 0) {
+    const currentLineItem = cart.lineItems.find((lineItem) => lineItem.productId === productId)
+    return currentLineItem ? currentLineItem.id : ''
+  } else {
+    return ''
+  }
+}
+
+export function extractProductPrices(product: Product, currentCurrencyCode: TCurrencyType) {
+  const prices = product.masterData.current.masterVariant.prices
+  let selectedPrice = {}
+  if (prices) {
+    prices.forEach((priceData) => {
+      if (priceData.value.currencyCode === currentCurrencyCode) {
+        selectedPrice = priceData
+      }
+    })
+  }
+  return selectedPrice
+}
+
+export function extractSrc(product: Product) {
+  return product.masterData.current.masterVariant.images
+    ? product.masterData.current.masterVariant.images[0]?.url
+    : ''
+}
+
+export function extractProductDescription(product: Product, lang: TLangType) {
+  return product.masterData.current.description
+    ? product.masterData.current.description[lang]
+    : ''
+}
+
+export function extractProductAttributes(product: Product, lang: TLangType) {
+  const rawAttributes = product.masterData.current.masterVariant.attributes
+  if (rawAttributes) {
+    return rawAttributes.reduce((acc: Record<string, string>[], attribute) => {
+      if (attribute.name === 'productspec') {
+        let label = []
+        if (attribute.value.label) {
+          label = attribute.value.label[lang]
+            .split('-')
+            .filter((val: string) => val.trim() !== '')
+            .map((val: string) => val.trim())
+        } else if (attribute.value) {
+          label = attribute.value[lang]
+            .split('-')
+            .filter((val: string) => val.trim() !== '')
+            .map((val: string) => firstLetterUppercase(val.trim()))
+        }
+        label.forEach((labelItem: string) => {
+          const formattedAttr = {
+            name: attribute.name,
+            key: `${attribute.value.key}-${labelItem}` || '',
+            label: labelItem || ''
+          }
+          acc.push(formattedAttr)
+        })
+      } else if (attribute.name === 'color-filter') {
+        let label = ''
+        if (attribute.value.label) {
+          label = attribute.value.label[lang]
+        } else {
+          label = attribute.value[lang]
+        }
+        if (label.charAt(0) !== '#') {
+          const formattedAttr = {
+            name: attribute.name,
+            key: `${attribute.value.key}` || '',
+            label: label || ''
+          }
+          acc.push(formattedAttr)
+        }
+      } else if (attribute.name === 'finish') {
+        let label = ''
+        if (attribute.value.label) {
+          label = attribute.value.label[lang]
+        } else {
+          label = attribute.value[lang]
+        }
+        if (label.charAt(0) !== '#') {
+          const formattedAttr = {
+            name: attribute.name,
+            key: `${attribute.value.key}` || '',
+            label: label || ''
+          }
+          acc.push(formattedAttr)
+        }
+      } else if (attribute.name === 'finishlabel') {
+        let label = ''
+        if (attribute.value.label) {
+          label = attribute.value.label[lang]
+        } else {
+          label = attribute.value[lang]
+        }
+        if (label.charAt(0) !== '#') {
+          const formattedAttr = {
+            name: attribute.name,
+            key: `${attribute.value.key}` || '',
+            label: label || ''
+          }
+          acc.push(formattedAttr)
+        }
+      } else if (attribute.name === 'color') {
+        let label = ''
+        if (attribute.value.label) {
+          label = attribute.value.label[lang]
+        } else {
+          label = attribute.value[lang]
+        }
+        if (label.charAt(0) !== '#') {
+          const formattedAttr = {
+            name: attribute.name,
+            key: `${attribute.value.key}` || '',
+            label: label || ''
+          }
+          acc.push(formattedAttr)
+        }
+      } else {
+        acc.push({})
+      }
+      return acc.filter((value) => Object.keys(value).length > 0)
+    }, [])
+  } else {
+    return []
+  }
+}
